+++ src/Service/Admin/AddOnsFactory.php
@@ -19,7 +19,7 @@ class AddOnsFactory
         'wp-statistics-advanced-reporting' => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=adv-report',
         'wp-statistics-customization'      => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=customization',
         'wp-statistics-widgets'            => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=adv-widgets',
-        'wp-statistics-realtime-stats'     => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=real-time',
+        'wp-statistics-realtime-stats'     => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=realtime',
         'wp-statistics-mini-chart'         => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=mini-chart',
         'wp-statistics-rest-api'           => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=rest-api',
         'wp-statistics-data-plus'          => 'utm_source=wp-statistics&utm_medium=link&utm_campaign=dp',
@@ -38,6 +38,18 @@ class AddOnsFactory
 
     private static function getFromRemote()
     {
+        // Define a unique transient key
+        $transientKey = 'wp_statistics_addons';
+
+        // Try to get the cached data
+        $cachedData = get_transient($transientKey);
+
+        // If the cached data is found, return it
+        if ($cachedData !== false) {
+            return $cachedData;
+        }
+
+        // If not found, fetch the data from the remote source
         $addOnsRemoteUrl = WP_STATISTICS_SITE . '/wp-json/plugin/addons';
         $response        = wp_remote_get($addOnsRemoteUrl, ['timeout' => 35]);
 
@@ -52,8 +64,13 @@ class AddOnsFactory
         $response = json_decode($response['body']);
 
         if (isset($response->items)) {
+            // Cache the data for 1 week
+            set_transient($transientKey, $response->items, WEEK_IN_SECONDS);
+
             return $response->items;
         }
+
+        return [];
     }
 
     public static function getSettingNameByKey($key)
@@ -66,5 +83,10 @@ class AddOnsFactory
     public static function getLicenseTransientKey($key)
     {
         return $key . '_license_response';
+    }
+
+    public static function getDownloadTransientKey($key)
+    {
+        return $key . '_download_info';
     }
 }